+++ packages/kit/test/apps/basics/test/client.test.js
@@ -1754,6 +1754,16 @@ test.describe('remote functions', () => {
 		expect(request_count).toBe(1); // no query refreshes, since that happens as part of the command response
 	});
 
+	test('query/command inside endpoint works', async ({ page }) => {
+		await page.goto('/remote/server-endpoint');
+
+		await page.getByRole('button', { name: 'get' }).click();
+		await expect(page.locator('p')).toHaveText('get');
+
+		await page.getByRole('button', { name: 'post' }).click();
+		await expect(page.locator('p')).toHaveText('post');
+	});
+
 	test('form.enhance works', async ({ page }) => {
 		await page.goto('/remote/form');
 		await page.fill('#input-task-enhance', 'abort');
@@ -1941,5 +1951,22 @@ test.describe('remote functions', () => {
 		// Verify pending count returns to 0
 		await expect(page.locator('#form-pending')).toHaveText('Form pending: 0');
 		await expect(page.locator('#form-button-pending')).toHaveText('Button pending: 0');
+	});
+
+	// TODO once we have async SSR adjust the test and move this into test.js
+	test('query.batch works', async ({ page }) => {
+		await page.goto('/remote/batch');
+
+		await expect(page.locator('#batch-result-1')).toHaveText('Buy groceries');
+		await expect(page.locator('#batch-result-2')).toHaveText('Walk the dog');
+		await expect(page.locator('#batch-result-3')).toHaveText('Buy groceries');
+		await expect(page.locator('#batch-result-4')).toHaveText('Error loading todo error: Not found');
+
+		let request_count = 0;
+		page.on('request', (r) => (request_count += r.url().includes('/_app/remote') ? 1 : 0));
+
+		await page.click('button');
+		await page.waitForTimeout(100); // allow all requests to finish
+		expect(request_count).toBe(1);
 	});
 });